pyproject.toml
[project]
name = "carebridge-langgraph-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi[standard]>=0.115.12",
    "googlemaps>=4.10.0",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.10",
    "langchain-tavily>=0.1.5",
    "langgraph>=0.3.20",
    "langsmith>=0.3.42",
    "python-dotenv>=1.1.0",
    "python-socketio>=5.13.0",
    "python-engineio>=4.0.0,<5.0.0",
    "uvicorn>=0.34.2",
    "spacy>=3.8.7",
]
